pyproject.toml
[project]
name = "find_upload"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-firestore>=2.22.0",
]

[project.scripts]
find_upload = "find_upload.cli:main"
